### 2019-08-01 #62 @labarba
👋 @magsol — You are now assigned as editor! Isn't our cute bot `whedon` the best? Your mission now is finding two willing reviewers for the submission. We have a list of folks who have signed up to review for JOSE, but you are free to seek outside of that, and invite via email, Twitter, or even a mention here on this thread.

We're working to update the JOSE website and docs, but in the meantime, the [Editorial Guide](https://joss.readthedocs.io/en/latest/editing.html) of our sister journal, JOSS, has all that you need to know.
