### 2019-09-21 #70 @AstrobioMike
Hi all,

In my opinion, this is all set to go. Things can be used on the hosted site perfectly as explained in the readme and nicely demonstrated in the [youtube video](https://www.youtube.com/watch?v=izER-V9PTxs&feature=youtu.be).

At the end of the readme there is a section on [how to host your own](https://github.com/Atticus29/population-fragmentation#how-to-set-up-your-own-instance-of-dragglesimulator-in-amazon-web-services) through amazon web services (AWS), which I had some trouble with. But that seems to me to be a completely extra, optional avenue, and not part of the core material or site that most people would use, so I don't think that should be an impediment to publication in JOSE. There are a lot of AWS-specific hurdles in the mix there like setting up the S3 bucket and setting up your IAM credentials – which the [readme notes](https://github.com/Atticus29/population-fragmentation#how-to-set-up-your-own-instance-of-dragglesimulator-in-amazon-web-services) in point #5 is demonstrated in the above linked video for setting up an S3 bucket, but is not. I needed to create an IAM user, which involved making a group and setting permissions and such which was not straightforward. After getting that, I also ran into issues with `ng build`, which led to a web of dependency and version problems related to angular that I eventually gave up on.

Again, everything works fine on the already-hosted site, and this section just seems to be completely optional information out of kindness, there for those who might want to set up their own. I don't see this as a core part of the educational program being offered here, and therefore don't think it should affect its publication in JOSE.

Regarding specific boxes left unchecked above by me as of this comment:

I left Version unchecked because the repo doesn't list any releases yet, maybe @Atticus29 is waiting for the review process to be done.

I left Installation and Installation Instructions left unchecked because there really is no installation to use the main platform that is already hosted. So i don't think these boxes are relevant, let me know if i should just check them if that's the case.

Minor notes for @Atticus29:

- Maybe add definition of “Population fragmentation” to the nice definitions page you have. And in there define “fragments” as used here, and its relation to subpopulations. For instance on the Create a Population page, it says to specify the number of fragments, but I wasn’t sure what was meant by that at first.

- In line with that, maybe there is room for more continuity in usage of “fragments” vs “subpopulations”. For instance on the Answer a Few Questions page, it has “Fragment Number” in the table as a column, but asks questions about “subpopulation”. There may be more nuanced reasons you’re using these the way you are that I’m just naive to, in which case of course ignore the suggestion :)

Cool site, @Atticus29!
