### 2019-09-25 #70 @wallacelab
I agree that this seems pretty much ready to go. End-user use is pretty straightforward and well documented. I didn't attempt the AWS installation because of the issues pointed out by @AstrobioMike, but the other aspects seem solid. (I was able to use custom Google sheets & forms, as advertised.)

Specific boxes I left unchecked:

- Version number: Not included for the reasons @AstrobioMike mentions
- Automatic tests: No automatic tests have been implemented according to the README
- References: The paper.md file seems to be lacking the reference to Hardy-Weinberg 1908. (It has an in-line citation but nothing under "references")

Other minor notes for potential improvement:
- On the definitions page, the last two terms ("sampling error" and "expected heterozygosity") have much longer, more involved definitions that don't seem to fit with the pattern of the others. ("Sampling error" especially was jarring since it takes most of the paragraph to get to the definition.) I'd recommend giving each of them a simple, succinct definition like the others, and put the longer explanation as a sub-bullet underneath.
- On the Draggle images, the spots seem a bit small and hard to see; I'd recommend making them much larger (see example image), using brighter shades of blue (cyan?) and green, and/or maybe making them thick stripes instead of spots. (Basically, just making them glaringly obvious.)
![draggle](https://user-images.githubusercontent.com/14350494/65637198-c80cd380-dfb1-11e9-816c-1210b531469c.png)
(The stripes may be overkill, but you get the idea.)
