### 2019-09-29 e24969b
update some fragment vs subpopulation phrasing per JOSE review
