+++ src/app/instructions/instructions.component.ts
@@ -16,7 +16,7 @@ import { DatabaseService } from '../database.service';
 export class InstructionsComponent implements OnInit {
   private introGoogleSheetUrl: String;
   private heading: String = "Instructions";
-  private instructions: String[] = ["Practice allele frequency calculations with your instructor", "Practice genotype frequency calculations with your instructor", "Practice expected heterozygosity calculations with your instructor", "Work through this simulation module, making sure to TRACK ALLELE FREQUENCIES OF EVERY SUBPOPULATION/FRAGMENT for EVERY generation","We'll start with 10 generations and the default allele frequencies", "Enter calculated allele frequencies in the spreadsheet", "Use allele frequecies to calculate expected genotype frequencies and expected heterozygosity", "Discuss class results", "Answer online worksheet questions"];
+  private instructions: String[] = ["Practice allele frequency calculations with your instructor", "Practice genotype frequency calculations with your instructor", "Practice expected heterozygosity calculations with your instructor", "Work through this simulation module, making sure to TRACK ALLELE FREQUENCIES OF EVERY SUBPOPULATION for EVERY generation","We'll start with 10 generations and the default allele frequencies", "Enter calculated allele frequencies in the spreadsheet", "Use allele frequecies to calculate expected genotype frequencies and expected heterozygosity", "Discuss class results", "Answer online worksheet questions"];
   private ngUnsubscribe: Subject<void> = new Subject<void>();
   private loading: boolean = true;
 
