### 2019-09-29 3712ded
make definitions in vocab component shorter and spots much larger on the draggles
