### 2019-09-29 #70 @Atticus29
Thanks so much for the feedback so far, @wallacelab and @AstrobioMike! Useful and much appreciated!
Today, I changed some of the wording re: fragement vs. subpopulation, shortened some of the definitions, and increased the radius of the spots by >=2x. It does look better that way, I think.
I also released v.1.0.0 in the repo.
