### 2019-09-29 7883f6d
change blue to cyan for better contrast
