### 2019-09-29 b9e6302
hopefully improve syntax in paper.bib
