+++ paper.md
@@ -26,7 +26,7 @@ bibliography: paper.bib
 
 The concept of genetic drift in population genetics is a notoriously difficult concept to teach and to learn. DraggleSimulator is an open-source simulation web application written using Google's Angular framework that allows students to create a metapopulation (a collection of isolated subpopulations) of draggles: fictitious, diploid (meaning they have a mom and a dad and two copies of every gene in their genomes, just like humans), monogamous (only mating with one other individual), and simultaneously-hermaphroditic animals (any individual can mate with any other individual and the sex of the individual is a non-issue). The animals mate at a rate consistent with replacement-level fertility (2 offspring per couple), and the simulator tracks these matings and how they change the composition of alleles (versions of genes) across generations.
 
-Upon creation, this metapopulation satisfies all assumptions of Hardy-Weinberg Equilibrium `[@Hardy:1908]` except infinite population size. These criteria only permit one force of evolutionary change to act: genetic drift, which is brought about by sampling error.
+Upon creation, this metapopulation satisfies all assumptions of Hardy-Weinberg Equilibrium [@Hardy08] except infinite population size. These criteria only permit one force of evolutionary change to act: genetic drift, which is brought about by sampling error.
 
 By experimenting in either a guided fashion as seen in this hosted instance of [Draggle Simulator](https://www.populationsimulator.com) or in a less-structured fashion, students can explore how allele frequencies can change in populations even in the absence of nonrandom mating, natural selection, mutation, or migration (the other better-known forces of evolutionary change). The students can compare how allele frequencies change in populations of different sizes or in metapopulations of the same total size but with different numbers of isolated subpopulations (or no isolated subpopulations at all). For instance, the simulation could be used to help students explore the effects of fragmenting populations (i.e., isolating into distinct subpopulations) on genetic diversity, which is what motivated me to write this software in the first place while teaching a conservation genetics unit in an environmental biology class.
 
