+++ paper.bib
@@ -1,4 +1,4 @@
-@online{DraggleSimulator,
+@Online{DraggleSimulator,
   author = {Mark Fisher},
   title = {DraggleSimulator: An Open Source Web Application for Teaching Genetic Drift: Source Code},
   year = 2019,
@@ -6,23 +6,14 @@
   urldate = {2019-07-15}
 }
 
-@online{DraggleSimulator,
-  author = {Mark Fisher},
-  title = {DraggleSimulator: An Open Source Web Application for Teaching Genetic Drift: Hosted Instance},
-  year = 2019,
-  url = {https://www.populationsimulator.com/},
-  urldate = {2019-07-15}
-}
-
-@article{10.1126/science.28.706.49,
-   author = {{Hardy}, G. H.
-	},
-    title = "{Mendelian proportions in a mixed population}",
-  journal = {\Science},
+@article{Hardy08,
+   author = {Hardy, G. H.},
+    title = {{Mendelian proportions in a mixed population}},
+  journal = {Science},
  keywords = {},
-     year = 1908,
-    month = july,
-   volume = 28,
+     year = {1908},
+    month = {July},
+   volume = {28},
     pages = {49-50},
       doi = {10.1126/science.28.706.49},
 }
