### 2019-09-29 #70 @Atticus29
Attempting to see whether my syntax change in the paper.bib file fixes the citation issue?...
