### 2019-09-29 3773618
fix typo in paper.md
