+++ paper.md
@@ -32,7 +32,7 @@ By experimenting in either a guided fashion as seen in this hosted instance of [
 
 # Statement of Need
 
-There are already good tools available for teaching students about genetic drift using simulations online. The problem is that most of the good ones i) cost students or instructors money to use and/or ii) allow for no customizability on the part of the instructor in terms of what types of data s/he wants the students to collect and what types of questions s/he wants to ask the students. DraggleSimulator addresses both of these shortcomings by providing students and instructors with a web application that is free and can be customized with easily-graded questions generated by the instructor using Google Forms (or by allowing the instructor to take advantage of one the default question set that is already publicly available). For information on how to set up these customizable questions, please see instructions in the README file.
+There are already good tools available for teaching students about genetic drift using simulations online. The problem is that most of the good ones i) cost students or instructors money to use and/or ii) allow for no customizability on the part of the instructor in terms of what types of data s/he wants the students to collect and what types of questions s/he wants to ask the students. DraggleSimulator addresses both of these shortcomings by providing students and instructors with a web application that is free and can be customized with easily-graded questions generated by the instructor using Google Forms (or by allowing the instructor to take advantage of the default question set that is already publicly available). For information on how to set up these customizable questions, please see instructions in the README file.
 
 DraggleSimulator also emphasizes the discovery-based approach to learning activities that is so popular (and, I think, rightly so) these days by requiring students to click buttons to create offspring and designate pairs of individuals to mate, rather than simply having students click, "run" and watch the simulation take off without understanding the underlying dynamics of how the new allele frequencies are generated and calculated.
 
