### 2019-09-29 0515a2d
update README with local dev instance details
