### 2019-10-20 #70 @labarba
@magsol — what is the status of this review? I see all reviewers have some items missing from their checklist. Could you check with them?
