### 2019-10-21 #70 @Atticus29
> @magsol — what is the status of this review? I see all reviewers have some items missing from their checklist. Could you check with them?

Thanks for the follow up, @labarba! Should I wait for @luizirber to weigh in as the final reviewer, or continue responding to the reviews I have?
