### 2019-10-22 232ef22
add caveat in AWS hosting instructions
