### 2019-10-30 #70 @magsol
Ok, it's definitely time to move forward on this.

First, can @AstrobioMike and @wallacelab please **confirm** your recommendations in favor of acceptance?
