### 2019-10-30 #70 @magsol
The article proof looks good and reads well to me. References and DOIs are in place.
