### 2019-10-30 #70 @magsol
Awesome, thank you @AstrobioMike and @wallacelab!

@Atticus29: would you please make a tagged release and archive, and post both in this thread? Keeping it at 1.0 is fine, just let me know if that's what you decide.
