### 2019-10-31 #70 @Atticus29
@magsol awesome! Thanks! I just released a new version (1.0.1): https://github.com/Atticus29/population-fragmentation/releases/tag/v1.0.1
Created an archive in zenodo:
http://doi.org/10.5281/zenodo.3523488

Hopefully it's ok that I'm the one doing this?:
