### 2019-11-03 #70 @labarba
@AstrobioMike, @wallacelab — I see you recommended publication, but your checklists have a couple of unchecked items. Are you OK checking those off?
