### 2019-11-03 #70 @AstrobioMike
I don’t think the installation boxes are relevant here (detailed in response [here](https://github.com/openjournals/jose-reviews/issues/70#issuecomment-533836711). Should I just check them anyway?
