### 2019-11-07 #70 @magsol
@labarba Looks like we're all checked off.
