### 2019-11-07 #70 @labarba
Congratulations, @Atticus29, your paper is published! 🚀

Heartfelt thanks to our editor: @magsol, and reviewers: @AstrobioMike, @wallacelab — your contribution to JOSE is greatly appreciated 🙏
