### 2019-11-07 #70 @labarba
https://twitter.com/JOSE_TheOJ/status/1192514412747341825
